Recalculate AEs and FNNs
