+++ dfpl/settings.py
@@ -42,7 +42,7 @@ nn_multi_target_numpy_type = np.float32
 # Training settings
 
 # Training settings of the AC that were magic numbers in the code before.
-ac_train_min_delta = 0.0001
+ac_train_min_delta = 0.001
 ac_train_check_period = 10
 ac_train_patience = 100
